Crypto-service changes around ECB decryption to work with MariaDB AES_ENCRYPT.
NOT READY FOR PRODUCTION USE; fuzzing of ECB fails and the MariaDB padding
schema is unknown. Also needs to handle "passphrases" in addition to raw keys.

1. Change from string to []byte; was initially thought to be simpler on the PHP
side, but caused a lot of headaches in Go. Can use base64 in the PHP client.

2. Add ECB decrypt support with padding to differentiate from proto-marshalled
GCM. TL;DR ECB will always be 16n bytes for integer n because the block size is
128 bits. The proto-marshalled GCM is padded to never be a multiple of 16 bytes.

3. Re-encryption now occurs for either key _or_ algorithm rotation.